+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 *Revisions*
 * Added a warning for when files aren't found. ([#29](https://github.com/hjwylde/git-fmt/issues/29))
 * Updated internal use of `git ls-files` to use the null terminator option. ([#27](https://github.com/hjwylde/git-fmt/issues/27))
+* Fixed debug log messages to have timestamp and log level on all lines. ([#33](https://github.com/hjwylde/git-fmt/issues/33))
 
 #### v0.1.0.2